# Optimizations to KMeans and IVFBuilder