--- a/Cargo.toml
+++ b/Cargo.toml
@@ -58,3 +58,4 @@ rayon = "1.10.0"
 sorted-vec = "0.8.5"
 dashmap = "6.1.0"
 reqwest = {version = "0.12.11", features = ["json"]}
+atomic_refcell = "0.1.13"