--- a/rs/index/Cargo.toml
+++ b/rs/index/Cargo.toml
@@ -22,3 +22,5 @@ tempdir.workspace = true
 utils.workspace = true
 serde.workspace = true
 serde_json.workspace = true
+rayon.workspace = true
+atomic_refcell.workspace = true