--- a/rs/index/src/spann/reader.rs
+++ b/rs/index/src/spann/reader.rs
@@ -55,7 +55,7 @@ mod tests {
             max_iteration: 1000,
             batch_size: 4,
             num_clusters,
-            num_data_points: num_vectors,
+            num_data_points_for_clustering: num_vectors,
             max_clusters_per_vector: 1,
             distance_threshold: 0.1,
             base_directory: base_directory.clone(),