## Summary
This adds 2 main implementation
* Parallelize posting list & kmeans builder, since they are independent and can be done in parallel
* SIMD the centroid computation

## Result
Before this change, for query-optimized configs (max posting list size = 2000)
* Index time: 96s
* Query latency: 3ms

After this change, for the same config:
* Index time: 56s
* Query latency: 800 us (<-- no this is not a typo)